Move more into chemistry.jl, and update tutorials to work outside VSCode
