+++ src/atmosphere.jl
@@ -1606,269 +1606,10 @@ module atmosphere
     end  # end of allocate
 
     """
-    **Normalise gas VMRs, keeping condensates unchanged**
-
-    Only acts on a single model level.
-
-    Parameters:
-    - `atmos::atmosphere.Atmos_t`       atmosphere structure
-    - `i::Int`                          level index to normalise
-    """
-    function normalise_vmrs!(atmos::atmosphere.Atmos_t, i::Int)
-        # Work variables
-        x_con::Float64 =    0.0
-        x_dry::Float64 =    0.0
-        x_dry_old::Float64= 0.0
-
-        # Work out total VMR of all condensing gases
-        for c in atmos.condensates
-            if atmos.gas_sat[c][i]
-                x_con += atmos.gas_vmr[c][i]
-            end
-        end
-
-        # Calculate current and target dry VMRs
-        x_dry = 1.0 - x_con
-        x_dry_old = 0.0
-        for g in atmos.gas_names
-            # skip condensing gases, since their VMR is set by saturation
-            if !atmos.gas_sat[g][i]
-                x_dry_old += atmos.gas_vmr[g][i]
-            end
-        end
-
-        # Renormalise VMR to unity, scaling DRY COMPONENTS ONLY
-        for g in atmos.gas_names
-            if !atmos.gas_sat[g][i]
-                atmos.gas_vmr[g][i] *= x_dry / x_dry_old
-            end
-        end
-
-        # Check total VMR at this level
-        x_tot = 0.0
-        for g in atmos.gas_names
-            x_tot += atmos.gas_vmr[g][i]
-        end
-        if abs(x_tot - 1.0) > 1.0e-5
-            @warn @sprintf("Mixing ratios sum to %.6e (level %d)",x_tot,i)
-        end
-
-        return nothing
-    end
-
-    """
-    **Adjust gas VMRs according to saturation and cold-trap requirements**
-
-    Volatiles which are allowed to condense are rained-out at condensing levels
-    until the gas is exactly saturated, not supersaturated. If evaporation is enabled here,
-    it will lead to enhanced mixing ratios at deeper levels as rain is converted back
-    into gas from a liquid state.
-
-    Arguments:
-    - `atmos::Atmos_t`          the atmosphere struct instance to be used.
-    """
-    function handle_saturation!(atmos::atmosphere.Atmos_t)
-
-        # Single gas case does not apply here
-        if atmos.gas_num == 1
-            return
-        end
-
-        # Parameters
-        evap_enabled::Bool =        false   # Enable re-vaporation of rain
-        evap_efficiency::Float64 =  0.5     # Evaporation efficiency
-
-        # Work arrays
-        maxvmr::Dict{String, Float64} = Dict{String, Float64}() # max running VMR for each condensable
-        cond_kg::Dict{String,Float64} = Dict{String, Float64}() # condensed kg/m2 for each condensable
-        x_sat::Float64 = 0.0
-        supcrit::Bool =  false
-
-        # Set maximum value (for cold trapping)
-        for c in atmos.condensates
-            maxvmr[c] = atmos.gas_vmr[c][end]
-        end
-
-        # Reset mixing ratios to surface values
-        # Reset phase change flags
-        # Reset condensation yield values
-        for g in atmos.gas_names
-           fill!(atmos.gas_vmr[g][1:end-1], atmos.gas_vmr[g][end])
-           fill!(atmos.gas_sat[g],          false)
-           fill!(atmos.gas_yield[g],        0.0)
-        end
-
-        # Reset water cloud
-        fill!(atmos.cloud_arr_r, 0.0)
-        fill!(atmos.cloud_arr_l, 0.0)
-        fill!(atmos.cloud_arr_f, 0.0)
-
-        # Handle condensation
-        for i in range(start=atmos.nlev_c-1, stop=1, step=-1)
-
-            # For each condensate
-            for c in atmos.condensates
-
-                # Reset condensation and rain
-                cond_kg[c] = 0.0   # kg/m2 of 'c' condensate produced at this level
-
-                # check criticality
-                supcrit = atmos.tmp[i] > atmos.gas_dat[c].T_crit+1.0e-5
-
-                # saturation mixing ratio
-                x_sat = phys.get_Psat(atmos.gas_dat[c], atmos.tmp[i]) / atmos.p[i]
-
-                # cold trap
-                if atmos.gas_vmr[c][i] > maxvmr[c]
-                    atmos.gas_vmr[c][i] = maxvmr[c]
-                    atmos.gas_sat[c][i] = true
-                end
-
-                # condense if supersaturated
-                if (atmos.gas_vmr[c][i] > x_sat) && !supcrit
-
-                    # set rainout kg/m2
-                    cond_kg[c] = atmos.gas_dat[c].mmw*atmos.p[i]*
-                                            (atmos.gas_vmr[c][i] - x_sat)/
-                                            (atmos.layer_grav[i] * atmos.layer_μ[i])
-
-                    # condensation yield at this level
-                    atmos.gas_yield[c][i] += cond_kg[c]
-
-                    # set new vmr
-                    atmos.gas_vmr[c][i] = x_sat
-
-                    # store vmr for cold trapping at levels above this one
-                    maxvmr[c] = x_sat
-
-                    # flag condensate as actively condensing at this level
-                    atmos.gas_sat[c][i] = true
-                    # @printf("%d: %s rain generated %.3e \n", i, c, cond_kg[c])
-
-                    # Set water cloud at this level
-                    if c == "H2O"
-                        # mass mixing ratio (take ratio of mass surface densities [kg/m^2])
-                        atmos.cloud_arr_l[i] = (cond_kg["H2O"] * atmos.cond_alpha) /
-                                                    atmos.layer_mass[i]
-
-                        if atmos.cloud_arr_l[i] > 1.0
-                            @warn "Water cloud mass mixing ratio is greater than unity (level $i)"
-                        end
-
-                        # droplet radius and area fraction (fixed values)
-                        atmos.cloud_arr_r[i] = atmos.cloud_val_r
-                        atmos.cloud_arr_f[i] = atmos.cloud_val_f
-                    end
-                end # end saturation check
-
-            end # end condensate
-
-            normalise_vmrs!(atmos, i)
-        end # end i levels
-
-        # recalculate layer properties
-        calc_layer_props!(atmos)
-
-        # Evaporate rain within unsaturated layers
-        if evap_enabled
-
-            total_rain::Float64 = 0.0
-            i_top_dry::Int = 1
-            i_bot_dry::Int = atmos.nlev_c
-
-            # For each condensable
-            for c in atmos.condensates
-
-                # reset dry region
-                i_top_dry = 1
-                i_bot_dry = atmos.nlev_c
-
-                # accumulate rain
-                total_rain = sum(atmos.gas_yield[c])
-
-                # no rain? go to next condensable
-                if total_rain < 1.0e-10
-                    continue
-                end
-
-                # locate top of dry region
-                for j in 1:atmos.nlev_c-1
-                    if atmos.gas_sat[c][j]
-                        # this layer is saturated => dry region must be below it
-                        i_top_dry = j+1
-                    end
-                end
-
-                # locate bottom of dry region
-                for j in range(start=atmos.nlev_c, stop=i_top_dry+1, step=-1)
-                    if atmos.tmp[j] > atmos.gas_dat[c].T_crit
-                        # this layer is supercritical => dry region must be above it
-                        i_bot_dry = j-1
-                    end
-                end
-                if i_bot_dry < i_top_dry
-                    i_bot_dry = i_top_dry
-                end
-
-                # evaporate rain in dry region
-                for j in i_top_dry+1:i_bot_dry
-
-                    # evaporate up to saturation
-                    dp_sat = phys.get_Psat(atmos.gas_dat[c], atmos.tmp[j]) -
-                                                    atmos.gas_vmr[c][j]*atmos.p[j]
-
-                    # Calculate kg/m2 of gas that would saturate layer j
-                    dm_sat = atmos.gas_dat[c].mmw * dp_sat/
-                                            (atmos.layer_grav[j] * atmos.layer_μ[j])
-
-                    # can we evaporate all rain within this layer?
-                    if total_rain < dm_sat
-                        # yes, so don't evaporate more rain than the total
-                        dm_sat = total_rain
-                    end
-                    atmos.gas_sat[c][j] = true
-
-                    # Evaporation efficiency factor
-                    #   This fraction of the rain that *could* be evaporated
-                    #   at this layer *is* converted to vapour in this layer.
-                    dm_sat *= evap_efficiency
-
-                    # offset condensate yield at this level by the evaporation
-                    atmos.gas_yield[c][j] -= dm_sat
-
-                    # add partial pressure from evaporation to pp at this level
-                    dp_sat = dm_sat * atmos.layer_grav[j] *
-                                              atmos.layer_μ[j] / atmos.gas_dat[c].mmw
-
-                    # convert extra pp to extra vmr
-                    atmos.gas_vmr[c][j] += dp_sat / atmos.p[j]
-
-                    # reduce total rain correspondingly
-                    total_rain -= dm_sat
-
-                    # Recalculate layer mmw
-                    atmos.layer_μ[j] = 0.0
-                    for g in atmos.gas_names
-                        atmos.layer_μ[j] += atmos.gas_vmr[g][j] * atmos.gas_dat[g].mmw
-                    end
-
-                end # go to next j level (below)
-
-            end # end loop over condensates
-
-        end # end evaporation scheme
-
-        # Recalculate layer properties
-        calc_layer_props!(atmos)
-
-        return nothing
-    end
-
-    """
     **Manually set water cloud properties at saturated levels.**
 
     Uses the default mass mixing ratio, area fraction, and droplet sizes.
-    This function is redundant if condensation is done via `handle_saturation!()`.
+    This function is redundant if condensation is done via `chemsitry.handle_saturation!()`.
 
     Arguments:
     - `atmos::Atmos_t`          the atmosphere struct instance to be used.
